16/6

+++ LiteCommerce.DomainModels/OrderDetail.cs
@@ -6,13 +6,15 @@ using System.Threading.Tasks;
 
 namespace LiteCommerce.DomainModels
 {
+    /// <summary>
+    /// Thông tin Chi tiết đơn hàng
+    /// </summary>
     public class OrderDetail
     {
         public int OrderID { get; set; }
         public int ProductID { get; set; }
-        public Int16 UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public int Quantity { get; set; }
         public int Discount { get; set; }
-
     }
 }
