+++ LiteCommerce.DomainModels/Product.cs
@@ -6,6 +6,9 @@ using System.Threading.Tasks;
 
 namespace LiteCommerce.DomainModels
 {
+    /// <summary>
+    /// Thông tin về Sản phẩm
+    /// </summary>
     public class Product
     {
         public int ProductID { get; set; }
@@ -13,9 +16,8 @@ namespace LiteCommerce.DomainModels
         public int SupplierID { get; set; }
         public int CategoryID { get; set; }
         public string QuantityPerUnit { get; set; }
-        public long UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public string Descriptions { get; set; }
         public string PhotoPath { get; set; }
-
     }
 }
